fix: move tests to another folder
